refactor: move modal utils out of dialogUtils.ts
