+++ Composer/packages/client/src/utils/constants.ts
@@ -1,0 +1,9 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+import { SDKKinds } from '@bfc/shared';
+
+export const eventTypeKey: string = SDKKinds.OnDialogEvent;
+export const intentTypeKey: string = SDKKinds.OnIntent;
+export const activityTypeKey: string = SDKKinds.OnActivity;
+export const customEventKey = 'OnCustomEvent';
